move simulation related code into its own namespace
